+++ PackagesOfFuture.WebUI/src/app/components/menu/menu.component.html
