+++ PackagesOfFuture.WebUI/src/app/shared.module.ts
@@ -1,7 +1,6 @@
 import { CommonModule } from '@angular/common'
 import { NgModule } from '@angular/core'
 import { FormsModule, ReactiveFormsModule } from '@angular/forms'
-import { LoginComponent } from './components/login/login.component'
 import { MaterialModule } from './material.module'
 
 const modules = [
@@ -11,10 +10,8 @@ const modules = [
   ReactiveFormsModule,
 ]
 
-const components = [LoginComponent]
 @NgModule({
-  declarations: [...components],
   imports: [...modules],
-  exports: [...modules, ...components],
+  exports: [...modules],
 })
 export class SharedModule {}
