--- PackagesOfFuture.WebUI/src/app/shared/services/claims/claims.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ClaimsServiceService } from './claims.service';
-
-describe('ClaimsServiceService', () => {
-  let service: ClaimsServiceService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ClaimsServiceService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
