--- PackagesOfFuture.WebUI/src/app/shared/services/claims/claims.service.ts
@@ -1,9 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class ClaimsServiceService {
-
-  constructor() { }
-}
