+++ PackagesOfFuture.WebUI/src/app/utilities/store-connected.component.ts
